validation-ui/requirements.txt
streamlit==1.29.0
neo4j==5.15.0
pandas==2.1.4
pydantic==2.5.0
python-dotenv==1.0.0
structlog==23.2.0
plotly==5.17.0
altair==5.2.0